\name{compass2DecimalDegree}
\alias{compass2DecimalDegree}
\title{compass2DecimalDegree}
\description{
Transforms the degrees from a sailors compass to the mathematical degrees
defined on the unit circle.
}
\usage{
compass2DecimalDegree(CompassDegree)
}
\arguments{
\item{CompassDegree}{Numeric number indicating the compass degree.}
}
\value{
\item{DecimalDegree}{Numeric number indicating the degree on the unit circle as
defined mathematically.}
}
\author{Quirin Stier}
