\name{computeSample4EM}
\alias{computeSample4EM}
\title{computeSample4EM}
\description{
computeSample4EM
}
\usage{
computeSample4EM(N, BigDataMaximum)
}
\arguments{
\item{N}{Integer meaning the number of observations.}
\item{BigDataMaximum}{Numeric between 0 and 1 meaning the percent of data to
use.}
}
\value{
\item{dataSampleIdx}{Numeric vector with sample of data for visualization.}
}
\author{Quirin Stier}
