\name{computeSample4Visualization}
\alias{computeSample4Visualization}
\title{computeSample4Visualization}
\description{
computeSample4Visualization
}
\usage{
computeSample4Visualization(N)
}
\arguments{
\item{N}{Integer meaning the number of observations.}
}
\value{
\item{dataSampleIdx}{Numeric vector with sample of data for visualization.}
}
\author{Quirin Stier}
