\name{decimal2CompassDegree}
\alias{decimal2CompassDegree}
\title{decimal2CompassDegree}
\description{
Transforms the degrees defined on the unit circle mathematically to the degrees
on the sailors compass.
}
\usage{
decimal2CompassDegree(DecimalDegree)
}
\arguments{
\item{DecimalDegree}{Numeric number indicating the degree on the unit circle as
defined mathematically.}
}
\value{
\item{CompassDegree}{Numeric number indicating the compass degree.}
}
\author{Quirin Stier}
